Check if service != null before calling method remove()

+++ js/lib/neard.js
@@ -200,7 +200,8 @@ neardService.unregisterService = function() {
 		}
 
 		function onAgentsRemoved(promise) {
-			neardService.service.remove().then(onSuccessCB, onErrorCB);				    	  
+			if (neardService.service != null)
+				neardService.service.remove().then(onSuccessCB, onErrorCB);				    	  
 		}
 			
 		function onErrorCB(error) {
